+++ pinot-query-runtime/src/main/java/org/apache/pinot/query/service/QueryServer.java
@@ -32,6 +32,7 @@ import org.apache.pinot.core.transport.grpc.GrpcQueryServer;
 import org.apache.pinot.query.runtime.QueryRunner;
 import org.apache.pinot.query.runtime.plan.DistributedStagePlan;
 import org.apache.pinot.query.runtime.plan.serde.QueryPlanSerDeUtils;
+import org.apache.pinot.query.service.dispatch.QueryDispatcher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
